backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "extruct>=0.18.0",
    "fastapi[standard]>=0.134.0",
    "google-genai>=1.65.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.37",
    "yt-dlp>=2026.2.21",
]
